requirements.txt
PyGithub==2.3.0
openai==1.50.0
python-dotenv==1.0.1
requests==2.32.3
